api_gateway/requirements.txt
Flask
requests
PyJWT
prometheus-flask-exporter
